iox-#1391 Move command line argument parser to 'cli' module
