+++ iceoryx_dust/test/moduletests/test_cli_command_line_parser.cpp
@@ -14,9 +14,9 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_dust/internal/cli/command_line_parser.hpp"
 #include "iceoryx_hoofs/error_handling/error_handling.hpp"
 #include "iceoryx_hoofs/testing/mocks/error_handler_mock.hpp"
+#include "iox/cli/command_line_parser.hpp"
 #include "iox/function.hpp"
 #include "iox/optional.hpp"
 #include "iox/std_string_support.hpp"
@@ -33,7 +33,6 @@ namespace
 {
 using namespace ::testing;
 using namespace iox::cli;
-using namespace iox::cli::internal;
 using namespace iox;
 using namespace iox::cxx;
